# docs/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/files": {
            "get": {
                "operationId": "FilesAgentController_findAll",
                "parameters": [
                    {
                        "name": "channel",
                        "required": true,
                        "in": "query",
                        "description": "Channel (agent) ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of attachments",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AttachmentResponse"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid API key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not your agent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ],
                "summary": "List all files in a channel",
                "tags": [
                    "Files"
                ]
            }
        },
        "/api/v1/files/upload": {
            "post": {
                "operationId": "FilesAgentController_upload",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": "Uploaded attachment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AttachmentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid API key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ],
                "summary": "Upload a file",
                "tags": [
                    "Files"
                ]
            }
        },
        "/api/v1/files/store": {
            "post": {
                "operationId": "FilesAgentController_store",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": "Stored attachment (orphan)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AttachmentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid API key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ],
                "summary": "Store a file without creating a message",
                "tags": [
                    "Files"
                ]
            }
        },
        "/api/v1/files/store-text": {
            "post": {
                "operationId": "FilesAgentController_storeText",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": "Stored text attachment (orphan)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AttachmentResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid API key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ],
                "summary": "Store text content as a file without creating a message",
                "tags": [
                    "Files"
                ]
            }
        },
        "/api/v1/files/{id}/download": {
            "get": {
                "operationId": "FilesAgentController_download",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File stream"
                    },
                    "401": {
                        "description": "Invalid API key",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Not your file",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Attachment not found",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api-key": []
                    }
                ],
                "summary": "Download file directly by attachment ID",
                "tags": [
                    "Files"
                ]
            }
        },
...
    },
    "info": {
        "title": "Placet API",
        "description": "Chat-based agent inbox for AI-human interaction",
        "version": "0.1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "http://localhost:3001",
            "description": "Local development"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            },
            "api-key": {
                "type": "apiKey",
                "in": "header",
                "name": "x-api-key",
                "description": "Placet API key (hp_...)"
            },
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "access_token"
            }
        },
        "schemas": {
            "ZodValidationError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "invalid_type"
                    },
                    "message": {
                        "type": "string",
                        "example": "Expected string, received number"
                    },
                    "path": {
                        "example": [
                            "email"
                        ],
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "code",
                    "message",
                    "path"
                ]
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "number",
                        "example": 400
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation failed"
                    },
                    "errors": {
                        "description": "Zod validation errors (only on 400)",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ZodValidationError"
                        }
                    }
                },
                "required": [
                    "statusCode",
                    "message"
                ]
            },
...
            "AttachmentResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "clxyz789"
                    },
                    "messageId": {
                        "type": "string",
                        "example": "clxyz123"
                    },
                    "pluginType": {
                        "type": "string",
                        "example": "file"
                    },
                    "filename": {
                        "type": "string",
                        "example": "report.pdf"
                    },
                    "mimeType": {
                        "type": "string",
                        "example": "application/pdf"
                    },
                    "size": {
                        "type": "number",
                        "example": 1024
                    },
                    "storageKey": {
                        "type": "string",
                        "example": "uploads/clxyz123/report.pdf"
                    },
                    "createdAt": {
                        "type": "string",
                        "example": "2025-01-01T00:00:00.000Z"
                    }
                },
                "required": [
                    "id",
                    "messageId",
                    "pluginType",
                    "filename",
                    "mimeType",
                    "size",
                    "storageKey",
                    "createdAt"
                ]
            },
...
        }
    }
}